# swagger/swagger.json
{
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "swagger": "2.0",
    "info": {
        "title": "SeqUI Server",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/seqapi/v1/async_search/start": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_start_async_search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.StartAsyncSearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.StartAsyncSearchResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/seqapi/v1/search": {
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "seqapi_v1"
                ],
                "operationId": "seqapi_v1_search",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Environment",
                        "name": "env",
                        "in": "query",
                        "required": true
                    },
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.SearchRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/seqapi.v1.SearchResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
...
        "/userprofile/v1/queries/favorite": {
            "get": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "userprofile_v1"
                ],
                "operationId": "userprofile_v1_getFavoriteQueries",
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/userprofile.v1.GetFavoriteQueriesResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "userprofile_v1"
                ],
                "operationId": "userprofile_v1_createFavoriteQuery",
                "parameters": [
                    {
                        "description": "Request body",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/userprofile.v1.CreateFavoriteQueryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response",
                        "schema": {
                            "$ref": "#/definitions/userprofile.v1.CreateFavoriteQueryResponse"
                        }
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        },
        "/userprofile/v1/queries/favorite/{id}": {
            "delete": {
                "security": [
                    {
                        "bearer": []
                    }
                ],
                "tags": [
                    "userprofile_v1"
                ],
                "operationId": "userprofile_v1_deleteFavoriteQuery",
                "parameters": [
                    {
                        "type": "string",
                        "format": "int64",
                        "description": "Favorite Query ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "A successful response"
                    },
                    "default": {
                        "description": "An unexpected error response",
                        "schema": {
                            "$ref": "#/definitions/UnexpectedError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "UnexpectedError": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                }
            }
        },
...
        "http.AsyncSearchRequestHistogram": {
            "type": "object",
            "properties": {
                "interval": {
                    "type": "string"
                }
            }
        },
...
        "seqapi.v1.Aggregation": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationBucket"
                    }
                },
                "not_exists": {
                    "type": "integer"
                }
            }
        },
        "seqapi.v1.AggregationBucket": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "not_exists": {
                    "type": "integer"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                },
                "value": {
                    "type": "number"
                }
            }
        },
        "seqapi.v1.AggregationFunc": {
            "type": "string",
            "enum": [
                "count",
                "sum",
                "min",
                "max",
                "avg",
                "quantile",
                "unique"
            ],
            "x-enum-varnames": [
                "afCount",
                "afSum",
                "afMin",
                "afMax",
                "afAvg",
                "afQuantile",
                "afUnique"
            ]
        },
        "seqapi.v1.AggregationQuery": {
            "type": "object",
            "properties": {
                "agg_func": {
                    "default": "count",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.AggregationFunc"
                        }
                    ]
                },
                "field": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "seqapi.v1.AggregationTsQuery": {
            "type": "object",
            "properties": {
                "agg_func": {
                    "default": "count",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.AggregationFunc"
                        }
                    ]
                },
                "field": {
                    "type": "string"
                },
                "group_by": {
                    "type": "string"
                },
                "interval": {
                    "type": "string",
                    "format": "duration",
                    "example": "1m"
                },
                "quantiles": {
                    "type": "array",
                    "items": {
                        "type": "number"
                    }
                }
            }
        },
...
        "seqapi.v1.Error": {
            "type": "object",
            "properties": {
                "code": {
                    "default": "ERROR_CODE_NO",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.ErrorCode"
                        }
                    ]
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "seqapi.v1.ErrorCode": {
            "type": "string",
            "enum": [
                "ERROR_CODE_NO",
                "ERROR_CODE_PARTIAL_RESPONSE",
                "ERROR_CODE_QUERY_TOO_HEAVY",
                "ERROR_CODE_TOO_MANY_FRACTIONS_HIT"
            ],
            "x-enum-varnames": [
                "aecNo",
                "aecPartialResponse",
                "aecQueryTooHeavy",
                "aecTooManyFractionsHit"
            ]
        },
        "seqapi.v1.Event": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "id": {
                    "type": "string"
                },
                "time": {
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
        "seqapi.v1.Histogram": {
            "type": "object",
            "properties": {
                "buckets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.HistogramBucket"
                    }
                }
            }
        },
        "seqapi.v1.HistogramBucket": {
            "type": "object",
            "properties": {
                "docCount": {
                    "type": "string",
                    "format": "uint64"
                },
                "key": {
                    "type": "string",
                    "format": "uint64"
                }
            }
        },
        "seqapi.v1.Order": {
            "type": "string",
            "enum": [
                "desc",
                "asc"
            ],
            "x-enum-varnames": [
                "oDESC",
                "oASC"
            ]
        },
        "seqapi.v1.SearchRequest": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationQuery"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "histogram": {
                    "type": "object",
                    "properties": {
                        "interval": {
                            "type": "string"
                        }
                    }
                },
                "limit": {
                    "type": "integer",
                    "format": "int32"
                },
                "offset": {
                    "type": "integer",
                    "format": "int32"
                },
                "order": {
                    "default": "desc",
                    "allOf": [
                        {
                            "$ref": "#/definitions/seqapi.v1.Order"
                        }
                    ]
                },
                "query": {
                    "type": "string"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "withTotal": {
                    "type": "boolean"
                }
            }
        },
        "seqapi.v1.SearchResponse": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Aggregation"
                    }
                },
                "error": {
                    "$ref": "#/definitions/seqapi.v1.Error"
                },
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.Event"
                    }
                },
                "histogram": {
                    "$ref": "#/definitions/seqapi.v1.Histogram"
                },
                "partialResponse": {
                    "type": "boolean"
                },
                "total": {
                    "type": "string",
                    "format": "int64"
                }
            }
        },
        "seqapi.v1.StartAsyncSearchRequest": {
            "type": "object",
            "properties": {
                "aggregations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/seqapi.v1.AggregationTsQuery"
                    }
                },
                "from": {
                    "type": "string",
                    "format": "date-time"
                },
                "histogram": {
                    "$ref": "#/definitions/http.AsyncSearchRequestHistogram"
                },
                "meta": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "retention": {
                    "type": "string",
                    "format": "duration",
                    "example": "1h"
                },
                "size": {
                    "type": "integer"
                },
                "to": {
                    "type": "string",
                    "format": "date-time"
                },
                "with_docs": {
                    "type": "boolean"
                }
            }
        },
        "seqapi.v1.StartAsyncSearchResponse": {
            "type": "object",
            "properties": {
                "search_id": {
                    "type": "string",
                    "format": "uuid"
                }
            }
        },
...
        "userprofile.v1.CreateFavoriteQueryRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "relativeFrom": {
                    "type": "string",
                    "format": "uint64"
                }
            }
        },
        "userprofile.v1.CreateFavoriteQueryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "int64"
                }
            }
        },
        "userprofile.v1.FavoriteQuery": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string",
                    "format": "int64"
                },
                "name": {
                    "type": "string"
                },
                "query": {
                    "type": "string"
                },
                "relativeFrom": {
                    "type": "string",
                    "format": "uint64"
                }
            }
        },
        "userprofile.v1.GetFavoriteQueriesResponse": {
            "type": "object",
            "properties": {
                "queries": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/userprofile.v1.FavoriteQuery"
                    }
                }
            }
        },
...
    },
    "securityDefinitions": {
        "bearer": {
            "description": "Authentication token, prefixed by Bearer: Bearer {token}",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}